@@ -14,17 +14,17 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>3.1.1</version>
+		<version>3.2.1</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 		<maven-jar-plugin.version>3.1.1</maven-jar-plugin.version>
-		<jakarta.annotation.api.version>2.1.0</jakarta.annotation.api.version>
-		<junit-jupiter.version>5.8.2</junit-jupiter.version>
+		<jakarta.annotation.api.version>2.1.1</jakarta.annotation.api.version>
+		<junit-jupiter.version>5.9.2</junit-jupiter.version>
 	</properties>
 
 	<dependencies>
@@ -67,7 +67,7 @@
 			<scope>test</scope>
 		</dependency>
 
-		<!-- Junit 5 -->
+		<!-- JUnit 5 -->
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
 			<artifactId>junit-jupiter-engine</artifactId>
@@ -94,6 +94,9 @@
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
+				<configuration>
+					<mainClass>com.in28minutes.spring.basics.SpringIn5StepsBasicApplication</mainClass>
+				</configuration>
 			</plugin>
 
 		</plugins>
