@@ -2,7 +2,7 @@ package com.in28minutes.mockito.mockitodemo.business;
 
 public class SomeBusinessImpl {
 	
-	private DataService dataService;
+	private final DataService dataService;
 	
 	public SomeBusinessImpl(DataService dataService) {
 		super();
@@ -23,6 +23,4 @@ public class SomeBusinessImpl {
 
 interface DataService {
 	int[] retrieveAllData();
-	
-	
 }
