@@ -15,10 +15,9 @@ public class SpringIn5StepsPropertiesApplication {
 
 	public static void main(String[] args) {
 
-		try (AnnotationConfigApplicationContext applicationContext = new AnnotationConfigApplicationContext(
-				SpringIn5StepsPropertiesApplication.class)) {
+		try (var applicationContext = new AnnotationConfigApplicationContext(SpringIn5StepsPropertiesApplication.class)) {
 
-			SomeExternalService service = applicationContext.getBean(SomeExternalService.class);
+			var service = applicationContext.getBean(SomeExternalService.class);
 			System.out.println(service.returnServiceURL());
 		}
 	}
