@@ -1,5 +1,5 @@
 package com.in28minutes.spring.basics.springin5steps.basic;
 
 public interface SortAlgorithm {
-	public int[] sort(int[] numbers);
+	int[] sort(int[] numbers);
 }
