@@ -2,16 +2,15 @@ package com.in28minutes.spring.basics.springin5steps;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-// replaced @RunWith with @ExtendWith
-// replaced SpringRunner.class with SpringExtension.class
 @ExtendWith(SpringExtension.class)
-//@SpringBootTest
-public class SpringIn5StepsBasicApplicationTests {
+@ContextConfiguration(locations="/testContext.xml")
+class SpringIn5StepsBasicApplicationTests {
 
 	@Test
-	public void contextLoads() {
+	void contextLoads() {
 	}
 
 }
