@@ -8,11 +8,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.context.annotation.ImportResource;
 
-// replaced @RunWith with @ExtendWith
-// replaced MockitoJUnitRunner.class with MockitoExtension.class
 @ExtendWith(MockitoExtension.class)
-public class SomeCdiBusinessTest {
+class SomeCdiBusinessTest {
 
 	// Inject Mock
 	@InjectMocks
@@ -23,19 +22,19 @@ public class SomeCdiBusinessTest {
 	SomeCdiDao daoMock;
 
 	@Test
-	public void testBasicScenario() {
+	void testBasicScenario() {
 		Mockito.when(daoMock.getData()).thenReturn(new int[] { 2, 4 });
 		assertEquals(4, business.findGreatest());
 	}
 
 	@Test
-	public void testBasicScenario_NoElements() {
+	void testBasicScenario_NoElements() {
 		Mockito.when(daoMock.getData()).thenReturn(new int[] { });
 		assertEquals(Integer.MIN_VALUE, business.findGreatest());
 	}
 
 	@Test
-	public void testBasicScenario_EqualElements() {
+	void testBasicScenario_EqualElements() {
 		Mockito.when(daoMock.getData()).thenReturn(new int[] { 2,2});
 		assertEquals(2, business.findGreatest());
 	}
